Server/requirements.txt
Flask
python-jose
flask_cors
flask_restful
python-dotenv
pymongo
Flask-Bcrypt